+++ example-usage/hasura/index.d.ts
@@ -25,15 +25,15 @@ export interface Client {
   wsClient?: SubscriptionClient
 
   query<R extends query_rootRequest>(
-    request: R,
+    request: R & { __name?: string },
   ): Promise<FieldsSelection<query_root, R>>
 
   mutation<R extends mutation_rootRequest>(
-    request: R,
+    request: R & { __name?: string },
   ): Promise<FieldsSelection<mutation_root, R>>
 
   subscription<R extends subscription_rootRequest>(
-    request: R,
+    request: R & { __name?: string },
   ): Observable<FieldsSelection<subscription_root, R>>
 
   chain: {
@@ -51,19 +51,19 @@ export type QueryResult<fields extends query_rootRequest> = FieldsSelection<
 >
 
 export declare const generateQueryOp: (
-  fields: query_rootRequest,
+  fields: query_rootRequest & { __name?: string },
 ) => GraphqlOperation
 export type MutationResult<
   fields extends mutation_rootRequest
 > = FieldsSelection<mutation_root, fields>
 
 export declare const generateMutationOp: (
-  fields: mutation_rootRequest,
+  fields: mutation_rootRequest & { __name?: string },
 ) => GraphqlOperation
 export type SubscriptionResult<
   fields extends subscription_rootRequest
 > = FieldsSelection<subscription_root, fields>
 
 export declare const generateSubscriptionOp: (
-  fields: subscription_rootRequest,
+  fields: subscription_rootRequest & { __name?: string },
 ) => GraphqlOperation